+++ test/test_optimize.py
@@ -25,8 +25,10 @@ class CmdStanMLETest(unittest.TestCase):
         for root, _, files in os.walk(DATAFILES_PATH):
             for filename in files:
                 _, ext = os.path.splitext(filename)
-                if ext.lower() in ('.o', '.d', '.hpp', '.exe', '') and (
-                    filename != ".gitignore"
+                if (
+                    ext.lower() in ('.o', '.d', '.hpp', '.exe', '')
+                    and filename != ".gitignore"
+                    and filename != "return_one.hpp"
                 ):
                     filepath = os.path.join(root, filename)
                     os.remove(filepath)
